chore: refactor integrations to packages
