--- apps/tanstack-trpc-app/src/integrations/todos/todos-trpc.ts
@@ -1,10 +1,0 @@
-import { createServerClient } from '@supabase/ssr';
-import { createTodolistsService } from '@workspace/todos-supabase/create-todolists-service';
-import { createTodosService } from '@workspace/todos-supabase/create-todos-service';
-
-export const createTodosTRPCContext = (
-  getSupabaseClient: ReturnType<typeof createServerClient>
-) => ({
-  todosService: createTodosService(getSupabaseClient),
-  todolistsService: createTodolistsService(getSupabaseClient),
-});
