+++ apps/tanstack-trpc-app/src/integrations/trpc/client.ts
@@ -1,5 +1,5 @@
 import { createTRPCClient } from '@workspace/trpc/client';
-import type { TRPCRouter } from './router';
+import type { TRPCRouter } from '@/integrations/trpc/router';
 
 export const createTRPCRouterClient = createTRPCClient<TRPCRouter>;
