+++ apps/tanstack-trpc-app/src/integrations/trpc/react.ts
@@ -1,4 +1,4 @@
-import { createTRPCContext } from '@trpc/tanstack-react-query';
 import type { TRPCRouter } from '@/integrations/trpc/router';
+import { createProvider } from '@workspace/tanstack-trpc/provider';
 
-export const { TRPCProvider, useTRPC } = createTRPCContext<TRPCRouter>();
+export const { TRPCProvider, useTRPC } = createProvider<TRPCRouter>();
