+++ packages/supabase/README.md
@@ -1,0 +1,8 @@
+# supabase
+
+This library was generated with [Nx](https://nx.dev).
+
+
+## Running unit tests
+
+Run `nx test supabase` to execute the unit tests via [Vitest](https://vitest.dev/).
