+++ packages/supabase/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@workspace/chat-trpc",
+  "name": "@workspace/supabase",
   "version": "0.0.1",
   "private": true,
   "type": "module",
@@ -15,11 +15,8 @@
     "vite-plugin-dts": "^4.5.3"
   },
   "dependencies": {
-    "zod": "^3.24.2",
     "@supabase/ssr": "^0.6.1",
     "@supabase/supabase-js": "^2.49.4",
-    "@workspace/chat-supabase": "workspace:*",
-    "@workspace/trpc": "workspace:*",
-    "@trpc/server": "^11.1.1"
+    "@tanstack/react-start": "^1.120.3"
   }
 }
