+++ packages/supabase/src/server-client.ts
@@ -1,7 +1,7 @@
 import { parseCookies, setCookie } from '@tanstack/react-start/server';
 import { createServerClient, parseCookieHeader } from '@supabase/ssr';
 
-export default function getSupabaseServerClient(request?: Request) {
+export function getSupabaseServerClient(request?: Request) {
   // Use provided cookies or get them from the current request
   const cookies = request
     ? parseCookieHeader(request?.headers.get('cookie') || request?.headers.get('Cookie') || '').reduce((acc, {name, value}) => ({
