+++ packages/supabase/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "../typescript-config/tsconfig.base.json",
+  "files": [],
+  "include": [],
+  "references": [
+    {
+      "path": "./tsconfig.lib.json"
+    },
+    {
+      "path": "./tsconfig.spec.json"
+    }
+  ],
+}
