+++ packages/supabase/tsconfig.lib.json
@@ -1,0 +1,29 @@
+{
+  "extends": "../typescript-config/tsconfig.lib.json",
+  "include": [
+    "src/**/*.ts"
+  ],
+  "paths": {
+    "@workspace/supabase/*": [
+      "./src/*"
+    ]
+  },
+  "references": [],
+  "compilerOptions": {
+    "lib": ["DOM", "DOM.Iterable", "ES2022", "ESNext"],
+    "outDir": "./out-tsc",
+    "tsBuildInfoFile": "./out-tsc/tsconfig.lib.tsbuildinfo"
+  },
+  "exclude": [
+    "vitest.config.ts",
+    "vitest.config.mts",
+    "src/**/*.test.ts",
+    "src/**/*.spec.ts",
+    "src/**/*.test.tsx",
+    "src/**/*.spec.tsx",
+    "src/**/*.test.js",
+    "src/**/*.spec.js",
+    "src/**/*.test.jsx",
+    "src/**/*.spec.jsx"
+  ]
+}
