+++ packages/tanstack-trpc/README.md
@@ -1,0 +1,9 @@
+# tanstack trpc 
+
+This library was generated with [Nx](https://nx.dev).
+
+
+
+## Running unit tests
+
+Run `nx test tanstack-trpc` to execute the unit tests via [Vitest](https://vitest.dev/).
