+++ packages/tanstack-trpc/eslint.config.mjs
@@ -1,0 +1,3 @@
+import eslint from '@workspace/eslint-config';
+
+export default [...eslint.configs.base, ...eslint.configs.lib];
