+++ packages/tanstack-trpc/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@workspace/tanstack-trpc",
+  "version": "0.0.1",
+  "private": true,
+  "type": "module",
+  "exports": {
+    "./query-client": "./src/query-client.ts",
+    "./context": "./src/context.tsx",
+    "./provider": "./src/provider.tsx"
+  },
+  "devDependencies": {
+    "@types/node": "^22.9.0",
+    "@types/react": "^19.0.10",
+    "@types/react-dom": "^19.0.0",
+    "@workspace/eslint-config": "workspace:*",
+    "@workspace/typescript-config": "workspace:*",
+    "jsonc-eslint-parser": "^2.4.0",
+    "vite": "^6.2.0",
+    "vite-plugin-dts": "^4.5.3"
+  },
+  "dependencies": {
+    "zod": "^3.24.2",
+    "superjson": "^1.0.0",
+    "@tanstack/react-query": "^5.76.0",
+    "@tanstack/react-query-devtools": "^5.76.0",
+    "@trpc/tanstack-react-query": "^11.1.1",
+    "@trpc/client": "^11.1.1",
+    "@trpc/server": "^11.1.1",
+    "@supabase/ssr": "^0.6.1",
+    "@supabase/supabase-js": "^2.49.4",
+    "@workspace/chat-supabase": "workspace:*",
+    "@workspace/trpc": "workspace:*",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
+  }
+}
