+++ packages/todos-trpc/package.json
@@ -16,6 +16,8 @@
   },
   "dependencies": {
     "zod": "^3.24.2",
+    "@supabase/ssr": "^0.6.1",
+    "@supabase/supabase-js": "^2.49.4",
     "@workspace/todos-supabase": "workspace:*",
     "@workspace/trpc": "workspace:*"
   }
