+++ packages/user-trpc/src/context.ts
@@ -1,5 +1,13 @@
 import { createServerClient } from "@supabase/ssr";
-import { createUserService } from "@workspace/user-supabase/create-user-service";
+import { createUserService, UserService } from "@workspace/user-supabase/create-user-service";
+
+export type Context = {
+  userService: UserService;
+  user: {
+    id: string;
+    email: string;
+  } | null;
+};
 
 export const createUserTRPCContext = (getSupabaseClient: ReturnType<typeof createServerClient>) => {
   const userService = createUserService(getSupabaseClient);
